requirements.txt
streamlit==1.34.0
argon2-cffi==21.3.0
pandas==2.0.3
scikit-learn==1.3.0
bokeh==2.4.3
numpy==1.26.2
openai
tenacity
python-dotenv
